feat: rv64 compilation without io

+++ crates/toolchain/platform/src/lib.rs
@@ -4,13 +4,13 @@
 #![deny(rustdoc::broken_intra_doc_links)]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
-#[cfg(target_os = "zkvm")]
+#[cfg(openvm_intrinsics)]
 pub use openvm_custom_insn::{custom_insn_i, custom_insn_r};
-#[cfg(target_os = "zkvm")]
+#[cfg(openvm_intrinsics)]
 pub mod alloc;
-#[cfg(all(feature = "rust-runtime", target_os = "zkvm"))]
+#[cfg(all(feature = "rust-runtime", openvm_intrinsics))]
 pub mod heap;
-#[cfg(all(feature = "export-libm", target_os = "zkvm"))]
+#[cfg(all(feature = "export-libm", openvm_intrinsics))]
 mod libm_extern;
 
 pub mod memory;
@@ -19,8 +19,8 @@ pub mod print;
 pub mod rust_rt;
 
 /// Size of a zkVM machine word in bytes.
-/// 4 bytes (i.e. 32 bits) as the zkVM is an implementation of the rv32im ISA.
-pub const WORD_SIZE: usize = core::mem::size_of::<u32>();
+/// 8 bytes (i.e. 64 bits) as the zkVM is an implementation of the rv64im ISA.
+pub const WORD_SIZE: usize = core::mem::size_of::<u64>();
 
 /// Standard IO file descriptors for use with sys_read and sys_write.
 pub mod fileno {
